backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development tools
ipython==8.17.2
django-extensions==3.2.3

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
django-stubs==4.2.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Debugging
django-debug-toolbar==4.2.0

